refactor: split ModList into smaller components for better maintainability
